+++ Configurations/PowerShellLab/VMConfiguration.ps1
@@ -24,7 +24,7 @@ Configuration AutoLab {
     Import-DSCResource -modulename "xActiveDirectory" -ModuleVersion  "3.0.0.0"
     Import-DSCResource -modulename "xComputerManagement" -ModuleVersion  "4.1.0.0"
     Import-DSCResource -modulename "xNetworking" -ModuleVersion  "5.7.0.0"
-    Import-DSCResource -modulename "xDhcpServer" -ModuleVersion  "2.0.0.0"
+    Import-DSCResource -modulename "xDhcpServer" -ModuleVersion  "3.0.0"
     Import-DSCResource -modulename 'xWindowsUpdate' -ModuleVersion  '2.8.0.0'
     Import-DSCResource -modulename 'xPendingReboot' -ModuleVersion  '0.4.0.0'
     Import-DSCResource -modulename 'xADCSDeployment' -ModuleVersion  '1.4.0.0'
@@ -49,7 +49,7 @@ Configuration AutoLab {
 
         registry TLS {
             Ensure = "present"
-            Key =  'HKLM:\SOFTWARE\Wow6432Node\Microsoft\.NetFramework\v4.0.30319' 
+            Key =  'HKLM:\SOFTWARE\Wow6432Node\Microsoft\.NetFramework\v4.0.30319'
             ValueName = 'SchUseStrongCrypto'
             ValueData = '1'
             ValueType = 'DWord'
@@ -272,6 +272,7 @@ Configuration AutoLab {
 
         xDhcpServerAuthorization 'DhcpServerAuthorization' {
             Ensure    = 'Present'
+            IsSingleInstance = 'yes'
             DependsOn = '[WindowsFeature]DHCP'
         }
 
@@ -287,6 +288,8 @@ Configuration AutoLab {
             DependsOn     = '[WindowsFeature]DHCP'
         }
 
+        <#
+        Deprecated
         xDhcpServerOption 'DhcpOption' {
             ScopeID            = $Node.DHCPScopeID
             DnsServerIPAddress = $Node.DHCPDnsServerIPAddress
@@ -294,6 +297,7 @@ Configuration AutoLab {
             AddressFamily      = $Node.DHCPAddressFamily
             DependsOn          = '[xDhcpServerScope]DhcpScope'
         }
+        #>
 
     } #end DHCP Config
     #endregion