+++ functions/public.ps1
@@ -77,6 +77,8 @@ Function Get-LabSummary {
 
     Begin {
         Write-Verbose "Starting $($myinvocation.mycommand)"
+        #create the media lookup table
+        lability\get-labmedia | foreach-object -begin { $media=[ordered]@{}} -Process { $media.Add($_.id,$_.description)}
     }
     Process {
         $Path = Convert-Path $path
@@ -85,44 +87,8 @@ Function Get-LabSummary {
 
         if (Test-Path $psd1) {
             $labname = Split-Path $Path -Leaf
-            Write-Verbose "getting summary for $labname"
+            Write-Verbose "Getting summary for $labname"
 
-            #could also use the Get-LabMedia command from the Lability module
-            $media = [ordered]@{
-                '2019_x64_Standard_EN_Eval'               = 'Windows Server 2019 Standard 64bit English Evaluation with Desktop Experience'
-                '2019_x64_Standard_EN_Core_Eval'          = 'Windows Server 2019 Standard 64bit English Evaluation'
-                '2019_x64_Datacenter_EN_Eval'             = 'Windows Server 2019 Datacenter 64bit English Evaluation with Desktop Experience'
-                '2019_x64_Datacenter_EN_Core_Eval'        = 'Windows Server 2019 Datacenter Evaluation in Core mode'
-                '2016_x64_Standard_EN_Eval'               = 'Windows Server 2016 Standard 64bit English Evaluation'
-                '2016_x64_Standard_Nano_DSC_EN_Eval'      = 'Windows Server 2016 Standard Nano 64bit English Evaluation'
-                '2016_x64_Standard_Core_EN_Eval'          = 'Windows Server 2016 Standard Core 64bit English Evaluation'
-                '2016_x64_Datacenter_EN_Eval'             = 'Windows Server 2016 Datacenter 64bit English Evaluation'
-                '2016_x64_Datacenter_Core_EN_Eval'        = 'Windows Server 2016 Datacenter Core 64bit English Evaluation'
-                '2016_x64_Standard_Nano_EN_Eval'          = 'Windows Server 2016 Standard Nano 64bit English Evaluation'
-                '2016_x64_Datacenter_Nano_EN_Eval'        = 'Windows Server 2016 Datacenter Nano 64bit English Evaluation'
-                '2012R2_x64_Standard_EN_Eval'             = 'Windows Server 2012 R2 Standard 64bit English Evaluation'
-                '2012R2_x64_Standard_EN_V5_Eval'          = 'Windows Server 2012 R2 Standard 64bit English Evaluation with WMF 5'
-                '2012R2_x64_Standard_EN_V5_1_Eval'        = 'Windows Server 2012 R2 Standard 64bit English Evaluation with WMF 5.1'
-                '2012R2_x64_Standard_Core_EN_Eval'        = 'Windows Server 2012 R2 Standard Core 64bit English Evaluation'
-                '2012R2_x64_Standard_Core_EN_V5_Eval'     = 'Windows Server 2012 R2 Standard Core 64bit English Evaluation with WMF 5'
-                '2012R2_x64_Standard_Core_EN_V5_1_Eval'   = 'Windows Server 2012 R2 Standard Core 64bit English Evaluation with WMF 5.1'
-                '2012R2_x64_Datacenter_EN_Eval'           = 'Windows Server 2012 R2 Datacenter 64bit English Evaluation'
-                '2012R2_x64_Datacenter_EN_V5_Eval'        = 'Windows Server 2012 R2 Datacenter 64bit English Evaluation with WMF 5'
-                '2012R2_x64_Datacenter_EN_V5_1_Eval'      = 'Windows Server 2012 R2 Datacenter 64bit English Evaluation with WMF 5.1'
-                '2012R2_x64_Datacenter_Core_EN_Eval'      = 'Windows Server 2012 R2 Datacenter Core 64bit English Evaluation'
-                '2012R2_x64_Datacenter_Core_EN_V5_Eval'   = 'Windows Server 2012 R2 Datacenter Core 64bit English Evaluation with WMF 5'
-                '2012R2_x64_Datacenter_Core_EN_V5_1_Eval' = 'Windows Server 2012 R2 Datacenter Core 64bit English Evaluation with WMF 5.1'
-                'WIN81_x64_Enterprise_EN_Eval'            = 'Windows 8.1 64bit Enterprise English Evaluation'
-                'WIN81_x64_Enterprise_EN_V5_Eval'         = 'Windows 8.1 64bit Enterprise English Evaluation with WMF 5'
-                'WIN81_x64_Enterprise_EN_V5_1_Eval'       = 'Windows 8.1 64bit Enterprise English Evaluation with WMF 5.1'
-                'WIN81_x86_Enterprise_EN_Eval'            = 'Windows 8.1 32bit Enterprise English Evaluation'
-                'WIN81_x86_Enterprise_EN_V5_Eval'         = 'Windows 8.1 32bit Enterprise English Evaluation with WMF 5'
-                'WIN81_x86_Enterprise_EN_V5_1_Eval'       = 'Windows 8.1 32bit Enterprise English Evaluation with WMF 5.1'
-                'WIN10_x64_Enterprise_EN_Eval'            = 'Windows 10 64bit Enterprise 1903 English Evaluation'
-                'WIN10_x86_Enterprise_EN_Eval'            = 'Windows 10 32bit Enterprise 1903 English Evaluation'
-                'WIN10_x64_Enterprise_LTSC_EN_Eval'       = 'Windows 10 64bit Enterprise LTSC 2019 English Evaluation'
-                'WIN10_x86_Enterprise_LTSC_EN_Eval'       = 'Windows 10 32bit Enterprise LTSC 2019 English Evaluation'
-            }
             Write-Verbose "Getting node data from $psd1"
             $import = Import-PowerShellDataFile -Path $psd1
             $Nodes = $import.allNodes